+++ R/add_attribute.R
@@ -7,20 +7,29 @@
 #'
 #' Users cannot assign names that are the same as the names of the core attributes. These functions are dedicated
 #' to adding data not part of the LAS specification. For example, \code{add_lasattribute(las, x, "R")}
-#' will fail because \code{R} is a name reserved for the red channel of las file that contains RGB attributes.\cr\cr
-#' \code{add_attribute} simply adds a new column in the data but does not update the header. Thus the LAS
+#' will fail because \code{R} is a name reserved for the red channel of las file that contains RGB attributes.
+#' Use \code{add_lasrgb} instead.
+#' \describe{
+#' \item{\code{add_attribute}}{Simply adds a new column in the data but does not update the header. Thus the LAS
 #' object is not strictly valid. These data will be temporarily usable at the R level but will not
-#' be written in a las file with \link{writeLAS}.\cr\cr
-#' \code{add_lasattribute} does the same as \code{add_attribute} but automatically updates the header of the
+#' be written in a las file with \link{writeLAS}.}
+#'
+#'\item{ \code{add_lasattribute}}{Does the same as \code{add_attribute} but automatically updates the header of the
 #' LAS object. Thus, the LAS object is valid and the new data is considered as "extra bytes". This new
-#' data will be written in a las file with \link{writeLAS}.\cr\cr
-#' \code{add_lasattribute_manual} allows the user to manually write all the extra bytes metadata.
+#' data will be written in a las file with \link{writeLAS}.}
+#'
+#' \item{\code{add_lasattribute_manual}}{Allows the user to manually write all the extra bytes metadata.
 #' This function is reserved for experienced users with a good knowledge of the LAS specifications.
-#' The function does not perform tests to check the validity of the information.\cr\cr
+#' The function does not perform tests to check the validity of the information.
 #' When using \code{add_lasattribute} and \code{add_lasattribute_manual}, \code{x} can only be of type numeric
 #' (\code{integer} or \code{double}). It cannot be of types \code{character} or \code{logical} as these are
 #' not supported by the las specifications. The types that are supported in lidR are types 0 to 10
-#' (table 24 page 25 of the specification). Types greater than 10 are not supported.
+#' (table 24 page 25 of the specification). Types greater than 10 are not supported.}
+#'
+#' \item{\code{add_lasrgb}}{Adds 3 columns named RGB and updates the point format of the LAS object
+#' for a format that supports RGB attributes. If the RGB values are ranging from 0 to 255 they are
+#' automatically scaled on 16 bits.}
+#' }
 #'
 #' @param las An object of class \link[lidR:LAS-class]{LAS}
 #' @param x a vector that needs to be added in the LAS object. For \code{add_lasattribute*} it can
@@ -31,6 +40,7 @@
 #' @param scale,offset numeric. The scale and offset of the data. NULL if not relevant.
 #' @param NA_value numeric or integer. NA is not a valid value in a las file. At time of writing it will
 #' be replaced by this value that will be considered as NA. NULL if not relevant.
+#' @param R,G,B integer. RGB values
 #'
 #' @return An object of class \link[lidR:LAS-class]{LAS}
 #'
@@ -155,6 +165,51 @@ add_lasattribute_manual = function(las, x, name, desc, type, offset = NULL, scal
   new_header <- rlas::header_add_extrabytes_manual(header, name, desc, type, offset, scale, NULL, NULL, NA_value)
   new_header <- LASheader(new_header)
   las@header <- new_header
+  return(las)
+}
+
+#' @export
+#' @rdname add_attribute
+add_lasrgb <- function(las, R, G, B)
+{
+  stopifnotlas(las)
+  stopifnot(is.integer(R), is.integer(G), is.integer(B))
+  assert_are_same_length(R, G)
+  assert_are_same_length(R, B)
+  assert_is_of_length(R, npoints(las))
+
+  maxr <- max(R, na.rm = TRUE)
+  maxg <- max(G, na.rm = TRUE)
+  maxb <- max(B, na.rm = TRUE)
+
+  scale <- 1L
+  if (maxr <= 255 & maxg <= 255 & maxb <= 255)
+    scale <- 257L
+
+  las@data$R <- R*scale
+  las@data$G <- G*scale
+  las@data$B <- B*scale
+
+  format <- las@header@PHB[["Point Data Format ID"]]
+
+  if (format %in% c(2,3,8))
+  {
+    # nothing to do
+  }
+  else if ("NIR" %in% names(las@data))
+  {
+    format <- 8L
+  }
+  else if ("gpstime" %in% names(las@data))
+  {
+    format <- 3L
+  }
+  else
+  {
+    format <- 2L
+  }
+
+  las@header@PHB[["Point Data Format ID"]] <- format
   return(las)
 }