+++ tests/testthat/test-lasadddata.R
@@ -4,6 +4,10 @@ LASfile <- system.file("extdata", "example.laz", package = "rlas")
 las     <- readLAS(LASfile, "xyz")
 x = runif(30)
 k = sample(1:100, 30)
+n = npoints(las)
+R = sample(1:255, n, TRUE)
+G = sample(1:255, n, TRUE)
+B = sample(1:255, n, TRUE)
 
 test_that("lasadddata works", {
    las2 = lasadddata(las, x, "plop")
@@ -36,5 +40,12 @@ test_that("lasremoveextrabyes works", {
   las3 = lasremoveextrabytes(las2, "plop")
 
   expect_equal(las3@header, las@header)
+})
+
+test_that("add_lasrgb works", {
+  las2 = add_lasrgb(las, R, G, B)
+
+  expect_equal(las2@header@PHB$`Point Data Format ID`, 2L)
+  expect_gt(mean(las2$R), 255)
 })